refactor: move applyLoanInterest to its own file
