+++ src/game-logic/reducers/applyLoanInterest.test.js
@@ -1,0 +1,12 @@
+import { applyLoanInterest } from './applyLoanInterest'
+
+describe('applyLoanInterest', () => {
+  test('applies loan interest', () => {
+    expect(
+      applyLoanInterest({
+        loanBalance: 100,
+        newDayNotifications: [],
+      }).loanBalance
+    ).toEqual(102)
+  })
+})
